feat: make App.js data root for notes
