+++ src/Notes/List.jsx
@@ -1,67 +1,32 @@
 /* eslint-disable no-underscore-dangle */
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import Note from './Note';
 
 const ListSt = styled.section`margin: 0 auto;`;
 
-class List extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      notes: [],
-    };
-    this.handleDelete = this.handleDelete.bind(this);
-  }
-
-  componentDidMount() {
-    this.props.fetchNotes().then((response) => {
-      this.setState({ notes: response.data });
-    });
-  }
-
-  handleDelete(noteIndex) {
-    const removeNote = (index) => {
-      this.setState(prevState => ({
-        notes: [...prevState.notes.slice(0, index), ...prevState.notes.slice(index + 1)],
-      }));
-    };
-    this.props.handleDelete(this.state.notes[noteIndex]._id).then(
-      () => {
-        removeNote(noteIndex);
-      },
-      () => {
-        alert('Ошибочка вышла');
-      },
-    );
-  }
-
-  render() {
-    return (
-      <div className="container">
-        <div className="row">
-          <div className="col-xs-12 col-md-8 col-md-offset-2">
-            <ListSt>
-              {this.state.notes.map((note, index) =>
-                (<Note
-                  title={note.title}
-                  text={note.text}
-                  id={index}
-                  key={note._id}
-                  handleDelete={this.handleDelete}
-                />),
-              )}
-            </ListSt>
-          </div>
-        </div>
+const List = props =>
+  (<div className="container">
+    <div className="row">
+      <div className="col-xs-12 col-md-8 col-md-offset-2">
+        <ListSt>
+          {props.notes.map((note, index) =>
+            (<Note
+              title={note.title}
+              text={note.text}
+              id={index}
+              key={note._id}
+              handleDelete={props.handleDelete}
+            />),
+          )}
+        </ListSt>
       </div>
-    );
-  }
-}
+    </div>
+  </div>);
 
 List.propTypes = {
-  fetchNotes: PropTypes.func.isRequired,
+  notes: PropTypes.array.isRequired,
   handleDelete: PropTypes.func.isRequired,
 };
